Fix vkb::screenshot()
See https://github.com/KhronosGroup/Vulkan-Samples/issues/259

As well as addressing the incorrect use of VK_IMAGE_USAGE_SAMPLED_BIT
as described in the issue, this change modifies the screenshot()
function to use VkCopyImageToBuffer rather than VkCopyImage as
it was using before. This results in both simpler and working
code (for me at least). The old version either produced empty
images, or images that started with a horizontal offset for me.